Add section for CPU Scaling
I asked a few people to show what profile their cpu was set to. All 4 of them showed theirs to be in powersave (With hosts from both Hetzner & OVH) and upon switching profile saw a SIGNIFICANT increase in clock speed and decrease in mspt. No other guide mentions this fact and i believe it has a major impact with no downsides.